+++ frontend-ds/gestor-proyecto/src/app/views/get_frontend/get_deportes.components.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import { AppService } from 'src/app/app.service';
+import swal from 'sweetalert2';
 
 @Component({
     selector: 'get_frontend',
@@ -31,7 +32,11 @@ export class GetDeporteComponent implements OnInit{
         this.service.get_deportes().subscribe(
             data => response = data,
             err => {
-                console.log("Error al consultar el servicio");
+                swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Algo salio mal!',
+                })
             },
             () => {
                 this.listado_deporte = response;
@@ -45,12 +50,22 @@ export class GetDeporteComponent implements OnInit{
         this.service.insert_deporte(this.deporte).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.deporte={
                 descripcion:""
             }
+            swal.fire({
+                title: 'Actividad deportiva agregada exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
         this.get_deportes();
         }
         );
@@ -62,15 +77,37 @@ export class GetDeporteComponent implements OnInit{
         var load={
             cod_act_deportiva:id
         }
-        this.service.delete_deporte(load).subscribe(
-        data => response = data,
-        err => {
-        console.log("Error al consultar servicio");
-        },
-        ()=> {
-            this.get_deportes();
-        }
-        );
+        swal.fire({
+            title: 'Advertencia!',
+            text: "Esta seguro desea borrar esta actividad deportiva?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+          }).then((result) => {
+            if (result.value) {
+                this.service.delete_deporte(load).subscribe(
+                    data => response = data,
+                    err => {
+                        swal.fire({
+                            icon: 'error',
+                            title: 'Oops...',
+                            text: 'Algo salio mal!',
+                        })
+                    },
+                    ()=> {
+                        this.get_deportes();
+                    }
+                    );
+              swal.fire(
+                'Borrado!',
+                'La actividad deportiva a sido borrada.',
+                'success'
+              )
+            }
+          })
     }
 
 
@@ -88,12 +125,22 @@ export class GetDeporteComponent implements OnInit{
         this.service.update_deporte(load, this.deporte).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.deporte = {
                 descripcion:""
             };
+            swal.fire({
+                title: 'Actividad deportiva actualizada exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
             this.get_deportes();
         }
         );
@@ -114,10 +161,4 @@ export class GetDeporteComponent implements OnInit{
     this.isShown = ! this.isShown;
 
     }
-
-    toggleShow1() {
-
-        this.isShown = ! this.isShown;
-    
-        }
 }
