+++ frontend-ds/gestor-proyecto/src/app/views/get_frontend/get_emocional.components.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import { AppService } from 'src/app/app.service';
+import swal from 'sweetalert2';
 
 @Component({
     selector: 'get_frontend',
@@ -30,7 +31,11 @@ export class GetEmocionalComponent implements OnInit{
         this.service.get_emocionales().subscribe(
             data => response = data,
             err => {
-                console.log("Error al consultar el servicio");
+                swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Algo salio mal!',
+                })
             },
             () => {
                 this.listado_emocional = response;
@@ -44,32 +49,68 @@ export class GetEmocionalComponent implements OnInit{
         this.service.insert_emocional(this.emocional).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.emocional={
                 descripcion:""
             }
+            swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Datos guardados correctamente',
+                showConfirmButton: false,
+                timer: 1500
+              })
         this.get_emocionales();
         }
         );
     }
 
     delete_emocional(id){
-        console.log("Registro a borrar: " + id);
         var response;
         var load={
             cod_problema_emocional:id
         }
-        this.service.delete_emocional(load).subscribe(
-        data => response = data,
-        err => {
-        console.log("Error al consultar servicio");
-        },
-        ()=> {
-            this.get_emocionales();
-        }
-        );
+        swal.fire({
+            title: 'Advertencia!',
+            text: "Esta seguro desea borrar este registro?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+          }).then((result) => {
+            if (result.value) {
+                this.service.delete_emocional(load).subscribe(
+                data => response = data,
+                err => {
+                    swal.fire({
+                        icon: 'error',
+                        title: 'Oops...',
+                        text: 'Algo salio mal!',
+                    })
+                },
+                ()=> {
+                    this.get_emocionales();
+                }
+                );
+            
+              swal.fire(
+                'Borrado!',
+                'El registro a sido borrado',
+                'success'                
+              )
+              
+            }
+          })
+        
+        
     }
 
 
@@ -86,13 +127,24 @@ export class GetEmocionalComponent implements OnInit{
         this.service.update_emocional(load, this.emocional).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.emocional = {
                 descripcion:""
             };
             this.get_emocionales();
+            swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Datos actualizados correctamente',
+                showConfirmButton: false,
+                timer: 1500
+              })
         }
         );
         
@@ -113,9 +165,4 @@ export class GetEmocionalComponent implements OnInit{
 
     }
 
-    toggleShow1() {
-
-        this.isShown = ! this.isShown;
-    
-        }
 }
