+++ frontend-ds/gestor-proyecto/src/app/views/get_frontend/get_frontend.components.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import { AppService } from 'src/app/app.service';
+import swal from 'sweetalert2';
 
 @Component({
     selector: 'get_frontend',
@@ -36,7 +37,11 @@ export class GetFrontendComponent implements OnInit{
         this.service.get_usuarios().subscribe(
             data => response = data,
             err => {
-                console.log("Error al consultar el servicio");
+                swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Algo salio mal!',
+                })
             },
             () => {
                 this.listado_info = response;
@@ -45,12 +50,24 @@ export class GetFrontendComponent implements OnInit{
         )
     }
 
+    errores(){
+        swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Algun campo requerido esta vacio!',
+        })
+    }
+
     insert_usuario(){
         var response;
         this.service.insert_usuario(this.usuario).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.usuario={
@@ -61,6 +78,12 @@ export class GetFrontendComponent implements OnInit{
                 correo: "",
                 tipo:""
             }
+            swal.fire({
+                title: 'Usuario agregado exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
         this.get_usuarios();
         }
         );
@@ -72,15 +95,38 @@ export class GetFrontendComponent implements OnInit{
         var load={
         codigo:id
         }
-        this.service.delete_usuario(load).subscribe(
-        data => response = data,
-        err => {
-        console.log("Error al consultar servicio");
-        },
-        ()=> {
-            this.get_usuarios();
-        }
-        );
+        swal.fire({
+            title: 'Advertencia!',
+            text: "Esta seguro desea borrar este usuario?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+          }).then((result) => {
+            if (result.value) {
+                this.service.delete_usuario(load).subscribe(
+                    data => response = data,
+                    err => {
+                        swal.fire({
+                            icon: 'error',
+                            title: 'Oops...',
+                            text: 'Algo salio mal!',
+                        })
+                    },
+                    ()=> {
+                        this.get_usuarios();
+                    }
+                    );
+              swal.fire(
+                'Borrado!',
+                'El usuario a sido borrado.',
+                'success'
+              )
+            }
+          })
+        
     }
 
 
@@ -97,7 +143,11 @@ export class GetFrontendComponent implements OnInit{
         this.service.update_usuario(load, this.usuario).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.usuario={
@@ -108,6 +158,12 @@ export class GetFrontendComponent implements OnInit{
                 correo: "",
                 tipo:""
             };
+            swal.fire({
+                title: 'Usuario actualizado exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
             this.get_usuarios();
         }
         );
